Split out base logic from LegacyHit into LegacyCirclePiece
